+++ src/App.js
@@ -1,12 +1,13 @@
 import React from 'react';
 import { Routes, Route } from 'react-router-dom';
 
-import { LandingPage } from 'pages';
+import { ErrorPage, HomePage } from 'pages';
 
 const App = () => {
     return (
         <Routes>
-            <Route path="/" exact element={<LandingPage />} />
+            <Route path="/" exact element={<HomePage />} />
+            <Route path="*" element={<ErrorPage />} />
         </Routes>
     );
 };