+++ src/components/common/Logo.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Logo = ({ className, ...props }) => {
-    return <></>;
-};
-
-export default Logo;