+++ src/components/index.js
@@ -1,3 +1,4 @@
-export { default as Logo } from 'components/common/Logo';
+export { default as BurguerMenu } from 'components/common/BurguerMenu';
+export { default as DoubleArrow } from 'components/common/DoubleArrow';
 export { default as Menu } from 'components/common/Menu';
 export { default as Navbar } from 'components/common/Navbar';