+++ src/index.js
@@ -16,6 +16,7 @@ const history = createBrowserHistory();
 ReactDOM.render(
     <AppContextProvider>
         <Router basename={parsePath()} history={history}>
+            <Navbar />
             <App />
         </Router>
     </AppContextProvider>,