+++ src/pages/home/Projects.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Projects = (props) => (
+    <div {...props} id="proyectos" className="h-screen bg-isabelline">
+        Proyectos
+    </div>
+);
+
+export default Projects;