+++ src/pages/landing-page/LandingPage.js
@@ -1,9 +1,0 @@
-const LandingPage = (props) => {
-    return (
-        <div {...props}>
-            <h2 className="font-normal text-4xl">Miguel Bruno</h2>
-        </div>
-    );
-};
-
-export default LandingPage;