+++ changelog.md
@@ -2,3 +2,4 @@
 
 -   Feature: Clone project from Vaikuntha.
 -   Feature: Remove unused code and add fonts.
+-   Feature: Add banner and bio sections.